game: re-request gamedata if needed
Sometimes on server reconnect, it seems gamedata is not sent (or goes
missing) on some games after re-connecting to them. This currently
results the game hanging until gtp2ogs is restarted and it gets another
chance to connect to the game.

As an immediate fix, re-request the gamedata if it has not been received
in 1 second.

It is not clear why the gamedata goes missing - there is no mention of
it on gtp2ogs side of logs so it is possible it happens due to something
on the server side.